ref: container the note on its own component for reusability
